Notes
Express enables to create multiple applications by sending that module into a variable "anyName"
Using that reference an application is created. It is useful if we have multiple web api on different ports
Example-
var Application1 = express();
Application2 = express();

To create router handlers or routes we have to use a "Router" that is an Express class. enables routing 
as well as 404 error handlers.

Body-parser parses the text as URL encoded data and shows the contained keys and values on request 
and response functions.

Ejs on the other hand is for using ejs files and tags i.e, embedded JavaScript, and by setting "view engine " 
as ejs  by a .set property so we  can render files which doesn't requires ejs extension while writing the code
example
res.render("Home.ejs"); This is if no view engine is declared.
res.render("Home"); This is if view engine is declared