Notes
An Html boilerplate can be divided into partials as a header and footer which can be used in our files
It is a concept which promotes DRY. We just have to include that header and footer inside ejs tags in 
those required files. Header and footer should be in partials dir and then partials in views dir.

To include header and footer inside a template we have to use ejs tags that are
"<% include partials/header %>" on the top of the template,
"<% include partials/footer  %>" on the bottom of the template.